Improved logging, added new `SpeedLimits` debug switch
Added a bunch of extra logging to the NetInfo scanner so we get more detail about any problems (will be useful to asset creators for example).

One of the log messages (that checks for applicalbe `m_netAI`) was spammy so added a new debug switch `SpeedLimits` to toggle that on and off.