Ensure a valid langauge is always used
This fixes an issue where user might select a new language, then revert to older version of mod that doesn't have that language; it would try using the previously selected language, not find it, and crash. The update ensures default language will be used in any case where selected language is not available.