Auto-debug ExtCitizenInstanceManager.StartPathFind()
If an exception occurs in `StartPathFind()` we don't get much information in the logs to help us determine what caused it.

This commit tries to remedy that by intercepting the exception then checking a `TernaryBool` (TB):

* `Undefined` = Set the TB to `True`
* `True` = Set the TB to `False`
* `False` = Do nothing

The original exception is then rethrown.

When the TB is `True`, the internal debug logging (all of it) is enabled for the next invocation of `StartPathFind()` which will dump all sorts of useful info to the `TMPE.log` file.

When an exception occurs, it's usually recurrent on subsequent calls to `StartPathFind()` so hopefully the debug logging will give us insight in to what specifically caused the problem.