Changed the GetNodeSegmentIds method to use a struct based iterator for performance reasons.
Provided some unit tests to see if that works properly.
Provided two benchmark tests to verify that the default implementation does not allocate on the heap if used directly while the indirect usage over the interface would cause boxing of the struct iterator.
Modified the benchmark program to use the BenchmarkSwitcher to be able to select the perf tests on start.