Minor UI adjustments. Simpler main window title
The version label doesn't have to be this prominent. Most users don't care about it during normal usage. On the few occasions where it is important, the user can easily look it up in their mod manager or Skyve.
The freed up screen space can now be used to spell out TM as Traffic Manager. This is clearer and avoids confusion, especially for new users. However I'd argue, that it even would make sense to get rid of the title label entirely.
Other than that, the "?"-button has been moved to the top-right corner and decreased in size a bit.